🐛 Fixed sending emails from email domain that includes www subdomain
fixes https://github.com/TryGhost/Team/issues/1855
fixes https://github.com/TryGhost/Team/issues/1866

Removed duplicate support address methods
